LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-alignment-implies-size-change-036-expected.txt



FAIL .before 1 assert_equals:
<img id="item" src="support/100x100-green.png" class=" before" data-expected-width="80" data-expected-height="80">
height expected 80 but got 100
PASS .after 2
